Updated stoichiometry to use integers

+++ trunk/psi-jami/jami-core/src/main/java/psidev/psi/mi/jami/model/Stoichiometry.java
@@ -15,12 +15,12 @@ public interface Stoichiometry {
      * will return the same mean stoichiometry value.
      * @return The minimum value for this stoichiometry if there is one, otherwise the mean stoichiometry value
      */
-    public long getMinValue();
+    public int getMinValue();
 
     /**
      * The maximum value for the stoichiometry. If the stoichiometry does not have a maxValue because it is a mean value, getMinValue() and getMaxValue()
      * will return the same mean stoichiometry value.
      * @return The maximum value for this stoichiometry if there is one, otherwise the mean stoichiometry value
      */
-    public long getMaxValue();
+    public int getMaxValue();
 }
